chore: address last circular references
