+++ src/classes/queue-base.ts
@@ -7,13 +7,13 @@ import {
 } from '../interfaces';
 
 import {
-  createScripts,
   delay,
   DELAY_TIME_5,
   isNotConnectionError,
   isRedisInstance,
   trace,
 } from '../utils';
+import { createScripts } from '../utils/create-scripts';
 import { RedisConnection } from './redis-connection';
 import { Job } from './job';
 import { KeysMap, QueueKeys } from './queue-keys';
