+++ src/utils/index.ts
@@ -9,17 +9,15 @@ import { CONNECTION_CLOSED_ERROR_MSG } from 'ioredis/built/utils';
 import {
   ChildMessage,
   ContextManager,
-  MinimalQueue,
   ParentOptions,
   RedisClient,
   Span,
   Tracer,
-} from './interfaces';
-import { Scripts } from './classes/scripts';
+} from '../interfaces';
 import { EventEmitter } from 'events';
 import * as semver from 'semver';
 
-import { SpanKind, TelemetryAttributes } from './enums';
+import { SpanKind, TelemetryAttributes } from '../enums';
 
 export const errorObject: { [index: string]: any } = { value: null };
 
@@ -233,22 +231,6 @@ export const childSend = (
   proc: NodeJS.Process,
   msg: ChildMessage,
 ): Promise<void> => asyncSend<NodeJS.Process>(proc, msg);
-
-/**
- * Factory method to create a Scripts object.
- */
-export const createScripts = (queue: MinimalQueue) => {
-  return new Scripts({
-    keys: queue.keys,
-    client: queue.client,
-    get redisVersion() {
-      return queue.redisVersion;
-    },
-    toKey: queue.toKey,
-    opts: queue.opts,
-    closing: queue.closing,
-  });
-};
 
 export const isRedisVersionLowerThan = (
   currentVersion: string,
